feat: represent compile time integer values as `i128` internally

+++ src/compiler/Ast.zig
@@ -82,7 +82,7 @@ pub const Node = union(enum) {
         };
 
         pub const Int = struct {
-            value: i64,
+            value: i128,
             source_loc: SourceLoc,
         };
 
@@ -455,7 +455,7 @@ pub const Parser = struct {
     }
 
     fn parseIntExpr(self: *Parser) Error!Node.Expr {
-        const value = std.fmt.parseInt(i64, self.tokenValue(self.peekToken()), 10) catch {
+        const value = std.fmt.parseInt(i128, self.tokenValue(self.peekToken()), 10) catch {
             self.error_info = .{ .message = "invalid number", .source_loc = self.tokenSourceLoc(self.peekToken()) };
 
             return error.InvalidNumber;
